Move config classes to sub-package
