+++ src/main/java/org/itsallcode/openfasttrace/Reporter.java
@@ -29,8 +29,8 @@ import java.util.List;
 import org.itsallcode.openfasttrace.core.Newline;
 import org.itsallcode.openfasttrace.core.Trace;
 import org.itsallcode.openfasttrace.FilterSettings;
-import org.itsallcode.openfasttrace.importer.legacytag.LegacyTagImporterConfig;
 import org.itsallcode.openfasttrace.importer.legacytag.LegacyTagImporterFactory;
+import org.itsallcode.openfasttrace.importer.legacytag.config.LegacyTagImporterConfig;
 import org.itsallcode.openfasttrace.report.ReportVerbosity;
 
 /**
