+++ src/test/java/org/itsallcode/openfasttrace/importer/legacytag/TestLegacyTagImporter.java
@@ -33,6 +33,7 @@ import java.nio.file.Paths;
 import org.itsallcode.openfasttrace.core.SpecificationItemId;
 import org.itsallcode.openfasttrace.importer.ImportEventListener;
 import org.itsallcode.openfasttrace.importer.LineReader;
+import org.itsallcode.openfasttrace.importer.legacytag.config.PathConfig;
 import org.junit.Before;
 import org.junit.Test;
 import org.mockito.InOrder;
