+++ src/test/java/org/itsallcode/openfasttrace/importer/legacytag/TestLegacyTagImporterFactory.java
@@ -37,6 +37,8 @@ import java.util.Optional;
 import org.itsallcode.openfasttrace.importer.ImportEventListener;
 import org.itsallcode.openfasttrace.importer.Importer;
 import org.itsallcode.openfasttrace.importer.ImporterException;
+import org.itsallcode.openfasttrace.importer.legacytag.config.LegacyTagImporterConfig;
+import org.itsallcode.openfasttrace.importer.legacytag.config.PathConfig;
 import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
