+++ src/test/java/org/itsallcode/openfasttrace/importer/legacytag/TestPathConfig.java
@@ -27,6 +27,7 @@ import static org.junit.Assert.assertThat;
 
 import java.nio.file.Paths;
 
+import org.itsallcode.openfasttrace.importer.legacytag.config.PathConfig;
 import org.junit.Test;
 
 public class TestPathConfig
